excercises 4.15 to 4.23

+++ part4/blog-list-backend/controllers/blogs.js
@@ -1,24 +1,33 @@
 const blogsRouter = require('express').Router()
+const middleware = require('../middlewares/middleware')
+
 const Blog = require('../models/blog')
 
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
-
 })
 
-blogsRouter.post('/',async (request, response) => {
-  const { title, url } = request.body
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const { title, author, url, likes } = request.body
+
   if(!title || !url) {
     return response.status(400).json({
       error: 'title or url missing',
     })
   }
 
-  const blog = new Blog(request.body)
+  const { user } = request
+  const blog = new Blog({ title, author, url, likes, user: user._id })
   blog.likes = blog.likes ?? 0
+
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+
   response.status(201).json(savedBlog)
 })
 
@@ -32,8 +41,18 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id',  middleware.userExtractor, async (request, response) => {
   const id = request.params.id
+  const { user } = request
+
+  const blog = await Blog.findById(id).populate('user', { username: 1, name: 1 })
+  if (!blog) {
+    return response.status(404).end()
+  }
+
+  if(blog.user._id.toString() !== user.id) {
+    return response.status(401).json({ error: 'unauthorized' })
+  }
 
   const deletedBlog = await Blog.findByIdAndDelete(id)
   if (!deletedBlog) {
@@ -43,7 +62,7 @@ blogsRouter.delete('/:id', async (request, response) => {
   return response.status(204).end()
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', middleware.userExtractor, async (request, response) => {
   const id = request.params.id
   const { title, author, url, likes  } = request.body
